Add files via upload
Before, this program is not separate via levels and the main file is too large, so I modify it to separate the levels and reduce the size of the main file.
